requirements.txt
asgiref==3.3.2
boto3==1.35.23
botocore==1.35.23
Django==3.2.22  # Upgraded Django version
djongo==1.3.6  # Replaces django-mongoengine
django-mongoengine
django-storages==1.14.4
djangorestframework==3.12.4  # Compatible with Django 3.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.8
inflection==0.5.1
jmespath==1.0.1
mongoengine==0.29.1
PyJWT==2.1.0  # Compatible with Django 3.2 and DRF
pytz==2024.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.2.4
typing-extensions==4.12.2
Werkzeug==1.0.1
python-decouple